+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/navigator/BuildTargetsActionGroup.java
@@ -231,7 +231,7 @@ public abstract class BuildTargetsActionGroup extends ViewPartActionGroup {
 						throws CommonException, OperationCancellation {
 					doJobRun(om);
 				}
-			}.executeAndHandle();
+			}.execute();
 		}
 		
 		protected String getJobTitle() {